Move to active_record_ext_7_0
